Converted the code to ARC (finally!)

+++ Desktop/Classes/LoggerUtils.m
@@ -36,8 +36,8 @@ NSString *StringWithTimeDelta(struct timeval *td)
 	{
 		int hrs,mn,s,ms;
 		hrs = (int)(td->tv_sec / 3600);
-		mn = (td->tv_sec % 3600) / 60;
-		s = td->tv_sec % 60;
+		mn = (int) ((td->tv_sec % 3600) / 60);
+		s = (int) (td->tv_sec % 60);
 		ms = td->tv_usec / 1000;
 		if (hrs != 0)
 			return [NSString stringWithFormat:@"+%dh %dmn %d.%03ds", hrs, mn, s, ms];
